+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## Unreleased
+## 4.0.0 - 2018-02-14
 
 #### Changes